Move USN specific enums to NTFS header file
