+++ tsk/fs/tsk_ntfs.h
@@ -605,9 +605,84 @@ extern "C" {
 
 #endif
 
-/* Update Sequence Journal Struct */
+/* Update Sequence Journal Structures */
 /************************************************************************
 */
+
+
+    enum TSK_FS_USN_REASON {
+        TSK_FS_USN_REASON_DATA_OVERWRITE = 0x00000001,
+        TSK_FS_USN_REASON_DATA_EXTEND = 0x00000002,
+        TSK_FS_USN_REASON_DATA_TRUNCATION = 0x00000004,
+        TSK_FS_USN_REASON_NAMED_DATA_OVERWRITE = 0x00000010,
+        TSK_FS_USN_REASON_NAMED_DATA_EXTEND = 0x00000020,
+        TSK_FS_USN_REASON_NAMED_DATA_TRUNCATION = 0x00000040,
+        TSK_FS_USN_REASON_FILE_CREATE = 0x00000100,
+        TSK_FS_USN_REASON_FILE_DELETE = 0x00000200,
+        TSK_FS_USN_REASON_EA_CHANGE = 0x00000400,
+        TSK_FS_USN_REASON_SECURITY_CHANGE = 0x00000800,
+        TSK_FS_USN_REASON_RENAME_OLD_NAME = 0x00001000,
+        TSK_FS_USN_REASON_RENAME_NEW_NAME = 0x00002000,
+        TSK_FS_USN_REASON_INDEXABLE_CHANGE = 0x00004000,
+        TSK_FS_USN_REASON_BASIC_INFO_CHANGE = 0x00008000,
+        TSK_FS_USN_REASON_HARD_LINK_CHANGE = 0x00010000,
+        TSK_FS_USN_REASON_COMPRESSION_CHANGE = 0x00020000,
+        TSK_FS_USN_REASON_ENCRYPTION_CHANGE = 0x00040000,
+        TSK_FS_USN_REASON_OBJECT_ID_CHANGE = 0x00080000,
+        TSK_FS_USN_REASON_REPARSE_POINT_CHANGE = 0x00100000,
+        TSK_FS_USN_REASON_STREAM_CHANGE = 0x00200000,
+        TSK_FS_USN_REASON_CLOSE = 0x80000000
+    };
+    typedef enum TSK_FS_USN_REASON TSK_FS_USN_REASON;
+
+
+    enum TSK_FS_USN_SOURCE_INFO {
+        TSK_FS_USN_SOURCE_INFO_DATA_MANAGEMENT = 0x01,
+        TSK_FS_USN_SOURCE_INFO_AUXILIARY_DATA = 0x02,
+        TSK_FS_USN_SOURCE_INFO_REPLICATION_MANAGEMENT = 0x04,
+        TSK_FS_USN_SOURCE_INFO_CLIENT_REPLICATION_MANAGEMENT = 0x08
+    };
+    typedef enum TSK_FS_USN_SOURCE_INFO TSK_FS_USN_SOURCE_INFO;
+
+
+    enum TSK_FS_NTFS_FILE_ATTRIBUTES {
+        TSK_FS_NTFS_FILE_ATTRIBUTE_READONLY = 0x000001,
+        TSK_FS_NTFS_FILE_ATTRIBUTE_HIDDEN = 0x000002,
+        TSK_FS_NTFS_FILE_ATTRIBUTE_SYSTEM = 0x000004,
+        TSK_FS_NTFS_FILE_ATTRIBUTE_DIRECTORY = 0x000010,
+        TSK_FS_NTFS_FILE_ATTRIBUTE_ARCHIVE = 0x000020,
+        TSK_FS_NTFS_FILE_ATTRIBUTE_DEVICE = 0x000040,
+        TSK_FS_NTFS_FILE_ATTRIBUTE_NORMAL = 0x000080,
+        TSK_FS_NTFS_FILE_ATTRIBUTE_TEMPORARY = 0x000100,
+        TSK_FS_NTFS_FILE_ATTRIBUTE_SPARSE_FILE = 0x000200,
+        TSK_FS_NTFS_FILE_ATTRIBUTE_REPARSE_POINT = 0x000400,
+        TSK_FS_NTFS_FILE_ATTRIBUTE_COMPRESSED = 0x000800,
+        TSK_FS_NTFS_FILE_ATTRIBUTE_OFFLINE = 0x001000,
+        TSK_FS_NTFS_FILE_ATTRIBUTE_NOT_CONTENT_INDEXED = 0x002000,
+        TSK_FS_NTFS_FILE_ATTRIBUTE_ENCRYPTED = 0x004000,
+        TSK_FS_NTFS_FILE_ATTRIBUTE_INTEGRITY_STREAM = 0x008000,
+        TSK_FS_NTFS_FILE_ATTRIBUTE_VIRTUAL = 0x010000,
+        TSK_FS_NTFS_FILE_ATTRIBUTE_NO_SCRUB_DATA = 0x020000
+    };
+    typedef enum TSK_FS_NTFS_FILE_ATTRIBUTES TSK_FS_NTFS_FILE_ATTRIBUTES;
+
+
+    typedef struct {
+        uint64_t refnum;
+        uint16_t refnum_seq;
+        uint64_t parent_refnum;
+        uint16_t parent_refnum_seq;
+        uint64_t usn;
+        uint32_t time_sec;
+        uint32_t time_nsec;
+        TSK_FS_USN_REASON reason;
+        TSK_FS_USN_SOURCE_INFO source_info;
+        uint32_t security;
+        TSK_FS_NTFS_FILE_ATTRIBUTES attributes;
+        char *fname;
+
+    } TSK_USN_RECORD_V2;
+
 
     typedef struct {
